added support for preset markers, some refectoring

+++ js/src/scripts.js
@@ -5,13 +5,13 @@
 var map = L.map('mapid', {
 	crs: L.CRS.Simple,
 	minZoom: 0,
-	maxZoom: 1,
+	maxZoom: 2,
 	zoomSnap: 0.5,
 	zoomDelta: 0.5,
 	maxBoundsViscosity: 0.8,
 });
 
-var bounds = [[-200,-200], [1744,1738]];
+var bounds = [[0,0], [1744,1738]];
 var image = L.imageOverlay('images/map.jpg', bounds).addTo(map);
 var markerData;
 var popupOptions;
@@ -35,7 +35,16 @@ window.onload = function() {
 
 // handle menu clicking
 $('.deleteAllMarkers:visible').click(function() {
-	deleteAllMarkers();
+	if(confirm('Are you sure you want to REMOVE all markers?')) {
+		deleteAllMarkers();
+		location.reload();
+	}
+});
+$('.resetMarkers:visible').click(function() {
+	if(confirm('Are you sure you want to RESET markers?')) {
+		addPresetMarkers();
+		location.reload();
+	}
 });
 
 // handle map clicking
@@ -54,6 +63,8 @@ function onMapClick(e) {
 	newMarker.bindPopup(popupContent, popupOptions).openPopup();
 	toolTipSettings.className = markerId;
 	newMarker.bindTooltip('0:00:00', toolTipSettings);
+
+	console.log(newMarker._latlng);
 }
 
 // handle events when popup is opened
@@ -138,7 +149,6 @@ function onReset(e) {
 
 function loadMarkers(markerObject) {
 	var newMarkerObject = Object();
-	console.log(markerObject);
 	for(var key in markerObject) {
 		var markerArray = markerObject[key];
 		var markerId = markerArray[0];
@@ -147,7 +157,6 @@ function loadMarkers(markerObject) {
 		var startDate = markerArray[4];
 		var markerType = markerArray[5];
 		var latlng = markerArray[1];
-		var popupContent = getPopupContent(markerId);
 
 		loadedMarker = new L.marker(latlng).addTo(map);
 		toolTipSettings.className = loadedMarker._leaflet_id;
@@ -155,6 +164,7 @@ function loadMarkers(markerObject) {
 
 		$('input[name="markerType' + loadedMarker._leaflet_id + '"][value="' + markerType + '"').prop('checked', true);
 
+		var popupContent = getPopupContent(loadedMarker._leaflet_id);
 		loadedMarker.on('popupopen', onPopupOpen);
 		loadedMarker.bindPopup(popupContent, popupOptions);
 
@@ -210,7 +220,7 @@ function getPopupContent(markerId) {
 	"<div class='marker-box'>" +
 		"<div class='marker-edit " + markerId + "'>" +
 			"<p>Time <input type='text' name='markerHours" + markerId + "' class='markerTextBox' id='markerHours' maxlength='1' /> : <input type='text' name='markerMins" + markerId + "' class='markerTextBox' id='markerMins' maxlength='2' /></p>" +
-			"<p><input type='radio' name='markerType" + markerId + "' value='lootCrate' /> Loot Crate<br /><input type='radio' name='markerType" + markerId + "' value='uplinkBlockade' /> Uplink/Blockade<br /><input type='radio' name='markerType" + markerId + "' value='custom' checked='checked' /> Custom</p>" +
+			"<p><label for='lootCrate'><input type='radio' name='markerType" + markerId + "' id='lootCrate' value='lootCrate' /> Loot Crate</label><br /><label for='uplinkBlockade'><input type='radio' name='markerType" + markerId + "' id='uplinkBlockade' value='uplinkBlockade' /> Uplink/Blockade<br /><label for='custom'><input type='radio' name='markerType" + markerId + "' id='custom' value='custom' checked='checked' /> Custom</label></p>" +
 			"<a href='#' class='saveMarkerButton'>Save</a> | <a href='#' class='deleteMarkerButton'>Delete</a> | <a href='#' class='resetMarkerButton'>Reset</a>" +
 	"</div>";
 
@@ -218,12 +228,91 @@ function getPopupContent(markerId) {
 }
 
 function deleteAllMarkers() {
-	if(confirm('Are you sure you want to delete all markers?')) {
-		localStorage.setItem('markerData', '');
-		location.reload();
-	}
+	localStorage.setItem('markerData', '');
+	return true;
 }
 
-function addLootCrates() {
-	
+function deletePresetMarkers() {
+	for(var key in markerData) {
+		var newData = markerData[key];
+		//console.log(newData);
+		if((newData[5] == "lootCrate") || (newData[5] == "uplinkBlockade")) {
+			delete markerData[key];
+		}
+	}
+
+	storeMarkers(markerData);
+}
+
+function addPresetMarkers() {
+	var startDate = Date.now();
+
+	// loot crates
+	var lootCratesObject = new Object();
+	lootCratesObject.silentValley = [368,1386];
+	lootCratesObject.summitAirfield = [1576,537];
+	lootCratesObject.summitAirfield2 = [1531,530];
+	lootCratesObject.summitAirfield3 = [1525,573];
+	lootCratesObject.sneakyDevilsberg = [1501,1000];
+	lootCratesObject.sneakyDevilsberg2 = [1480,930];
+	lootCratesObject.hellsPeak = [1321,681];
+	lootCratesObject.northernExcavation = [1251,1300];
+	lootCratesObject.hotelBravo = [965,1402];
+	lootCratesObject.wolfpineTunnel = [860,389];
+	lootCratesObject.wolfpine = [810,483];
+	lootCratesObject.mistyvale = [925,858];
+	lootCratesObject.mistyvale2 = [846,828];
+	lootCratesObject.mistyvale3 = [835,828];
+	lootCratesObject.mistyvale4 = [824,840];
+	lootCratesObject.mistyvale5 = [826,923];
+	lootCratesObject.patsRepairShop = [792,1058];
+	lootCratesObject.melsGrainFarm = [640,988];
+	lootCratesObject.southernExcavation = [449,679];
+	lootCratesObject.jaysResort = [484,943];
+	lootCratesObject.jaysResort2 = [474,924];
+	lootCratesObject.adamsScrapyard = [571,1110];
+	lootCratesObject.adamsScrapyard2 = [534,1100];
+	lootCratesObject.fortQCLewis = [593,1357];
+	lootCratesObject.fortQCLewis2 = [571,1359];
+
+	// blockades
+	var blockadesObject = new Object();
+	blockadesObject.westBlockade = [1179,622];
+	blockadesObject.eastBlockade = [756,1394];
+
+	// uplinks
+	var uplinksObject = new Object();
+	uplinksObject.westUplink = [1053,298];
+	uplinksObject.westUplink2 = [772,666];
+	uplinksObject.eastUplink = [973,1487];
+
+	deletePresetMarkers();
+
+	var genId = 0;
+	for(var key in lootCratesObject) {
+		var newLatLng = lootCratesObject[key];
+		var newData = ['lootCrate' + genId, newLatLng, 0, 0, startDate, 'lootCrate'];
+		markerData['lootCrate' + genId] = newData;
+		genId += 1;
+	}
+
+	var genId = 0;
+	for(var key in blockadesObject) {
+		var newLatLng = blockadesObject[key];
+		var newData = ['blockade' + genId, newLatLng, 0, 0, startDate, 'uplinkBlockade'];
+		markerData['blockade' + genId] = newData;
+		genId += 1;
+	}
+
+	var genId = 0;
+	for(var key in uplinksObject) {
+		var newLatLng = uplinksObject[key];
+		var newData = ['uplink' + genId, newLatLng, 0, 0, startDate, 'uplinkBlockade'];
+		markerData['uplink' + genId] = newData;
+		genId += 1;
+	}
+
+	storeMarkers(markerData);
+
+	return true;
 }
